feat: update header design + add css variables for colors
The header responsiveness wasn't consistent. Font-sizes were either
too big or too small. Decided it was best to design with a mobile first
approach and make use of the entire screen width. The cut off is 1300px
and any screen larger, the content would be cenetered on the screen with
a wrapper of max width 1300px.

- index.html:
  - add .innerWrapper div to contain header elements
  - this way at larger screens (1300px +) elements will be centered.
- header.css:
  - update styles for new design and better responsiveness.
  - desigining with smaller screens first and making changes on larger
  screens
  - use css variables for colors
- main.css:
  - update to use css variables for colors
  - inverse .innerWrapper class
- skills.css:
  - update color for css variable
- bio.js:
  - remove unused classes from HTML template strings (first/last name)