fix: jump effect on scroll for chrome/FF to using vh on mobile/tablet
On Chrome/Firefox, when scrolling there would be a jump effect.
Resulting in a bad UX.

The reason this happens because the hero uses vh unit to size itself.
On these mobile browsers, the menu bars (url + nav) disappears on scroll
causing all layout re-paint and re-adjust since the dimensions have
changed which results in a jump effect.

To fix this, decided it would be best to target these browsers and set a
fixed height to the hero to stop re-paint on scroll when the menu bars
appear/disappear.

More info:
https://stackoverflow.com/questions/37112218/css3-100vh-not-constant-in-mobile-browser

More on fix:
https://css-tricks.com/the-trick-to-viewport-units-on-mobile/

-index.html:
  - add script to check that we're on mobile/tablet and on Chrome/FF
    - if true, then calculate viewport height to fix hero height
    - does re-calculation if screen orientation changes
- header.css:
  - add height calculation to .hero where needed for mobile/tablet
- main.css:
  - add --vh css variable for fallback if js fails to run
  - remove min-height from body since it really isn't needed
    - was only there at the start when I needed to see full site layout
     before content was added