fix: project image heights not equalizing on page refresh
Sometimes on page refresh, project images did not equalize and
would be set to height 0.
This is expected, since on $(document).ready() we try to equalize
image heights and this doesn't mean images have loaded yet.

The fix was to wait until all content is loaded and then equalize the
project image heights.