fix: font re-render on page load
On page load, the font-family would re-render from the fallback font.
This happened because the Roboto font-family load was being called from
within the stylesheet.
So after the stylesheet loaded a call to the google font api was made.

The fix was to load the google font Roboto out of styles.css

- index.html:
  - add google font api stylesheet link
- main.css:
  - remove @import to google font api